Use NSInteger, not int, fixing a 64-to-32 warning.

+++ Core/Source/GrowlPreferencePane.m
@@ -750,7 +750,7 @@
 		[selectedNotificationIndexes release];
 		selectedNotificationIndexes = [newSelectedNotificationIndexes copy];
 
-		int indexOfMenuItem = [[notificationDisplayMenuButton menu] indexOfItemWithRepresentedObject:[[notificationsArrayController selection] valueForKey:@"displayPluginName"]];
+		NSInteger indexOfMenuItem = [[notificationDisplayMenuButton menu] indexOfItemWithRepresentedObject:[[notificationsArrayController selection] valueForKey:@"displayPluginName"]];
 		if (indexOfMenuItem < 0)
 			indexOfMenuItem = 0;
 		[notificationDisplayMenuButton selectItemAtIndex:indexOfMenuItem];
